First commit. This doesn't work yet!
Just to get something up so I stop stressing about code quality.
It will make the new-style SQL tables which are *incompatible with the web visualizations*.
I need to put the core fields that are part of the SAM specification in to the code (not as external modules), then develop a dependancy system for modules to only gather the bare minimum of states needed (in order) by having the python dynamically written, much like what was done for the SeQL project. Hopefully all done by tomorrow.